jdt.ls: add `classFileContentsSupport` capability
When working on a Java project that runs on either Maven or Gradle, jdt.ls takes
care of providing ycmd (and therefore all its clients) with neat and useful
features like autocompletion suggestions, documentation on the spot, etc.

One feature in particular does not appear to be working to its full potential:
going to the definition of a class that is defined or declared _outside_ of the
project, using some variation of the ':GoTo', either by typing that command
directly or using a shortcut that executes it.

Particulary, this is what you get greeted with once you attempt to go to the
definition of a class located outside of your project (and could be, for
example, located inside a JAR file as it is a project dependency):

    RuntimeError: Cannot jump to location

The primary reason is that ycmd is not sending any JSON related to extending the
capabilities of jdt.ls, and as a result, we are not getting any `jdt://` URIs
that point to the files where the "definitions" we want are located.

We receive nothing in return when the client makes a POST request to the
endpoint '/run_completer_command' due to the fact that jdt.ls believes that we
are not capable of handling `jdt://` URIs, so it spares us from any troubles.

Assuming we want support for such feature, we have to start from somewhere.
Particularly, we have to let jdt.ls know that we are now able to handle `jdt://`
URIs. Specifically, what needs to be sent is the following JSON during
initialization in order to deliver the good news:

    'extendedClientCapabilities': {
        'classFileContentsSupport': True
    }

Introduce a new function AddExtendedClientCapabilities() which takes in a
'settings' parameter and simply adds the _correct_ JSON necessary for jdt.ls to
start responding back with `jdt://` URIs. Existing settings provided in
.ycm_extra_conf.py are not touched, we simply "append" the above JSON to what
already exists.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>