jdt.ls: support jumping to `jdt://` URIs with GoTo
Great. Every time ycmd initializes with jdt.ls, it makes the server aware that
it is able to fetch Java class file contents and properly handle them
("classFileContentsSupport" extended capability) - but in reality, that's _not_
the case.

In order to make this work, change the handling of incoming 'jdt://` URIs that
jdt.ls currently fires off as needed.

The goal: use these `jdt://` URIs to jump to the definition of a class or method
located in a JAR file (for example, when checking the definition of, say, a
method or class from a library). Focus on GoTo() that is located in the file
`language_server_completer.py`, since executing `:YcmCompleter GoTo` should work
even when the class or method under the cursor is declared or defined inside the
Maven/Gradle project, not just in a dependency.

Introduce a function IsJdtContentUri() that takes a string input `uri` and
checks whether the first few characters match "jdt://".

Use this inside GoTo() in `language_server_completer.py` to determine whether to
fetch the contents behind the URI by sending another request to the language
server: `java/classFileContents` — the extended capability enabled by adding
`classFileContentsSupport` to the `extendedClientCapabilities` list.

When GoTo() receives a `jdt://` URI, send a request to get the file contents
(along with cursor data, specifically where to place the cursor), package all
that in a JSON, and hand it over to YouCompleteMe for display.

In that JSON, name the key for the file contents "jdt_contents" to communicate
to YouCompleteMe that these contents should be displayed in a temporary readonly
vim buffer.

The communication flow for jumping to `jdt://` URIs with GoTo:

    ycm                     ycmd                      jdt.ls
     |                       |                          |
     |----GoToDefinition---->|                          |
     |                       |                          |
     |                       |-textDocument/definition->|
     |                       |                          |
     |                       |<------LSP response-------|
     |                       |                          |
     |                       |-java/classFileContents-->|
     |                       |                          |
     |                       |<---class file contents---|
     |                       |                          |
     |<--jdt_contents JSON---|                          |
     |                       |                          |

A short description of each part of the flow, starting from the top (first) to
the bottom (last):

    1. This is the part where we run the command from our client, where we
    execute `:YcmCompleter GoToDefinition`.

    2. ycmd sends a `textDocument/definition` to get the `file://` or `jdt://`
    location and the range pointing to the definition location.

    3. If the langauge server responds with a `file://`, then it's business as
    usual. If a `jdt://` URI appears, send a `java/classFileContents` request
    [4] to retrieve the file contents.

    4. (If [3] returned a `jdt://`) Send `java/classFileContents` to the
    langauge server to retrieve the file contents.

    5. Language server responds with the class file contents.

    6. Package the file contents, range, and `jdt://` URI in a JSON and send it
    to ycm, the original requester. ycm should now make a temporary readonly
    vim buffer with the contents from the JSON and pointed at the cursor
    coordinates found in the JSON.

There are surely ways to send a better JSON payload back to ycm (the JSON in
step [6]), but this approach does the trick for now and handles other cases as
well.

Additionally, do not treat `jdt://` URIs as file paths, and for the places where
they risk being treated as so, conditionally return the file path itself with
the `jdt://` included, or if it's not a `jdt://` path, then run the normal
conditions.

In _LspSymbolListToGoTo(), filter the list of locations to not include any
JDT-related results so that they do not appear when typing `:YcmCompleter
GoToSymbol <symbol>`. It wouldn't work anyways as these file paths do not point
to an actual file and the default behavior is for the client to open an empty
new file with the name of the whole JDT URI - behavior that we do not desire.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>